refactor: HitterRecordDto 패키지 변경
