+++ src/main/kotlin/com/example/kbocombo/crawler/game/application/HitterRecordClient.kt
@@ -1,6 +1,7 @@
 package com.example.kbocombo.crawler.game.application
 
-import com.example.kbocombo.record.application.HitterRecordDto
+import com.example.kbocombo.crawler.game.infra.dto.HitterRecordDto
+
 
 interface HitterRecordClient {
