+++ src/main/kotlin/com/example/kbocombo/crawler/game/infra/dto/Dto.kt
@@ -1,0 +1,9 @@
+package com.example.kbocombo.crawler.game.infra.dto
+
+import com.example.kbocombo.player.vo.WebId
+
+data class HitterRecordDto(
+    val webId: WebId,
+    val atBats: Int,
+    val hits: Int
+)
